IMPALA-2885: ScannerContext::Stream objects should be owned by ScannerContext
Each scanner context has at least one stream which corresponds to
a scan range. For parquet scanner, there can be multiple streams.
These Stream objects are stored in the RuntimeState's object pool
even though they have the same life span of the scanner threads.
This change makes ScannerContext the owner of the Stream objects
so they will be freed when ScannerContext is destroyed.

Change-Id: Ic5440d414ecc0ca19676c553275aeb85231d6045
Reviewed-on: http://gerrit.cloudera.org:8080/3590
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Internal Jenkins