IMPALA-2700: ASCII NUL characters are doubled on insert into text tables
Currently the scanner processes the '\0' character as a no special character
whereas the writer treats it as a special character. The writer
appends a special character before writting which is causing the ASCII
NULL characters to double since they are the default escape characters.
This adds a check to treat '\0' as a no special character in the
writter.

Change-Id: Ia30fa314d1ee1e99f9e7598466eb1570ca7940fc
Reviewed-on: http://gerrit.cloudera.org:8080/3876
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins