IMPALA-3336: qgen: do not randomly generate query options (workaround)
Disabling random query options seems to lessen the effects of
IMPALA-3336. Since much of the memory for IMPALA-3336 is tied up in the
psycopg2 cursor, and these query options are for Impala, there must be
something going on in the query execution path. One theory is that so
many random values for these query options don't make sense, and more
queries are rejected, but meanwhile they continue to run against
Postgres. It's possible the Postgres threads are not cleaning up.

I haven't been able to dive into that, but let's disable this for now.
We're able to get coverage without it. Moreover, I think more care needs
to be given to the values chosen for these query options.

Change-Id: If9c10e98f3be19739e223e89ffd79880e4c31de2
Reviewed-on: http://gerrit.cloudera.org:8080/3712
Tested-by: Michael Brown <mikeb@cloudera.com>
Reviewed-by: David Knupp <dknupp@cloudera.com>