IMPALA-3225: Make a gerrit instance that can push to git.apache.org
This is an accompanying change to a change with the same JIRA ID in
Impala-auxiliary-tests. This change makes sure that no one can call
gerrit-verify-merge any longer and can only call gerrit-verify-only
once we completely switch development over to the new Apache Git repo.

This has to be merged only after the accompanying change has been
merged to Impala-axuiliary-tests.

Change-Id: Ia4f1135a536da9f6b9bf2720c31fcc005f2a1da4
Reviewed-on: http://gerrit.cloudera.org:8080/3742
Reviewed-by: Jim Apple <jbapple@cloudera.com>
Tested-by: Internal Jenkins