IMPALA-3225: Fix and update push_to_asf.py to match the new Impala-ASF project
The script now points to the updated 'Impala-ASF' repo on Gerrit.
Some fixes were also made to the error messages and how we handle
remotes.

Now we have the following options:
'--gerrit_remote' to set any gerrit remote name of the committers
choice.
'--apache_remote' also lets any remote name to point to the Apache
git repo.

The Gerrit URL is now hardcoded and no longer an option because the
Gerrit project name has been decided and is unlikely to change.

This is a follow up commit to other commits with the same JIRA ID.

Change-Id: Ieac2c2715ced99fd0ff3e3c1422187d5d1eef3d9
Reviewed-on: http://gerrit.cloudera.org:8080/3759
Reviewed-by: Jim Apple <jbapple@cloudera.com>
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Sailesh Mukil <sailesh@cloudera.com>