IMPALA-3857: KuduScanNode race on returning "optional" threads
The KuduScanNode could return all active scanner threads
when there were no more "optional" tokens available. In this
case, any remaining scan ranges wouldn't be picked up and
the query would produce incorrect results. This fixes the
issue by cleaning up the ScannerThread code and making sure
not to exit the last thread.

This was tested by running the tpch workload repeatedly
under load. That work to incorporate tpch data loading for
Kudu is actively being worked on, so this was tested manually.

Change-Id: I22adf2109b43b1b37d9a597de85e063431dff155
Reviewed-on: http://gerrit.cloudera.org:8080/3798
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins