IMPALA-3790: Fix admission control flaky stress test
In addition to a previous change which extended the
admission control test timeouts for code coverage jobs, the
tests with high concurrency are still experiencing timeouts
in the admission control queues (which is different from the
timeouts that were set on the test cases). Rather than
extend the timeouts on the queues as well (which would
increase the already ridiculously long test time ~2hrs),
this limits the number of concurrent queries that are
submitted with code coverage.

Change-Id: Id62f7603f1174aa02469c6ca57513c3f1fa1e221
Reviewed-on: http://gerrit.cloudera.org:8080/3861
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins