IMPALA-3957: Test failure in S3 build: TestLoadData.test_load
The test_load() test failed with an error which said that the number of
of files in the destination was wrong. This could probably be because
the filesystem_client.copy() (for S3) in the setup_method() silently
failed without copying one of the files as a one off error.

I'm not sure why S3 failed to do the copy, but this patch adds an
assert after the copy to make sure that if it the s3_client.copy()
fails to do the copy, it will assert instead of continuing with the
rest of the tests.

Change-Id: I966a469e94099d3d971e470ae6e992386070c5e9
Reviewed-on: http://gerrit.cloudera.org:8080/3881
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Internal Jenkins