IMPALA-3952: Clear scratch batch mem pool if Open() failed.
We used to be able to hit a DCHECK in HdfsParquetScanner::Close()
when using the legacy aggs/joins if HdfsParquetScanner::Open() failed.
With the legacy aggs/joins the tuple ptrs of the scratch batch are
allocated from the scratch batch's mem pool, and if Open() failed we
never freed or transferred the scratch batch's mem pool.

Testing: I tested this patch together with the fix for IMPALA-3964
on core/hdfs with the legacy aggs and joins enabled.

Change-Id: I55f32ed698a5b6fed8c28af1391aa07e1560e782
Reviewed-on: http://gerrit.cloudera.org:8080/3953
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins