IMPALA-3996: Migrate to updated Kudu insert string API
In Kudu 0.10 the Insert API changed for Strings and Binary
values to avoid ambiguity around memory copying. They
changed SetString() to copy the value given to it to avoid
misuse, and added a SetStringNoCopy() with the old behavior.

We upgraded to 0.10 which has some perf impact but is still
correct. This changes the code to use the NoCopy call which
was the behavior we had previously.

See Kudu commit:
https://github.com/apache/kudu/commit/48766a4ce17d422ced9a6ec78c9a9969ac44d8c9

Change-Id: I910c24724d0bc887b2d4a3e62ecdf72420a76f6f
Reviewed-on: http://gerrit.cloudera.org:8080/4055
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins