IMPALA-4009: qgen: add docs, utility for installing Oracle as ref database
This patch adds documentation and a small utility for users wishing to
run the random query generator against Oracle as a reference database.
In particular, the patch

* points users toward installing Oracle Database.
* provides directions on incorporating a Python Oracle client
  (cx_Oracle) into our impala-python virtual environment.
* provides a small utility that will verify installation and
  connectivity between client and server

This documentation isn't meant to be a complete guide. Other patches
will add documentation for actually loading random query generator test
data into Oracle, for example.

Change-Id: Ib3bf84b7afa8851c49a8d0f0a1ceded94f4de158
Reviewed-on: http://gerrit.cloudera.org:8080/4095
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Matthew Jacobs <mj@cloudera.com>