IMPALA-3567: move ExecOption profile helpers to RuntimeProfile
This is groundwork for IMPALA-3567, which will move some logic that
previously resided in ExecNodes into DataSinks. We want to report
ExecOption strings consistently in both ExecNodes and DataSinks,
so that logic needs to move to a shared place (e.g. the RuntimeProfile
itself).

I ran the patch through clang-format, which changed line wrapping and
indentation of a few surrounding lines.

Change-Id: I21c1dda8f8a1d92172bf59fbc1070a6834e61913
Reviewed-on: http://gerrit.cloudera.org:8080/4188
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins