IMPALA-4028: Improve message for improper Sentry config to make extra spaces visible.
When the Sentry config file contained extra spaces it used to be difficult to
determine what went wrong because the offending path was not quoted making the
spaces hard to see. As a result, Impala startup failed.

This patch quotes the bad file path in the error message to make the extra spaces
visible.

Change-Id: I3a76b9e4236caa3f2088fba8a9cf0236fced2634
Reviewed-on: http://gerrit.cloudera.org:8080/4309
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins