IMPALA-4087: TestFragmentLifecycle.test_failure_in_prepare
The test previously got the "initial" value of number of fragments by
reading the metric. This gave an incorrect non-zero result if there were
any leftover queries running on the cluster.

Avoid the problem and simplify the test by explicitly waiting for the
number of fragments to go to zero.

Change-Id: I112e502a25f075928b0f6ef376c7fd9c6376ef4d
Reviewed-on: http://gerrit.cloudera.org:8080/4325
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins