IMPALA-4104: add DCHECK to ConsumeLocal() and fix tests
The TestEnv used for the backend tests does not connect up the
MemTracker hierarchy in the expected way. This caused the valid
DCHECK in ConsumeLocal() to be triggered in backend tests.

This change fixes TestEnv to set up MemTrackers with the normal
hierarchy, as shown below, and fixes the tests to deal with the fallout
of that.

(Process)
 |
(Query)----------
 |              |
(Block Mgr) (Fragment instance)

Change-Id: Iadcbe96a9f1bf19872436211b049cebf39b0afe7
Reviewed-on: http://gerrit.cloudera.org:8080/4531
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins