IMPALA-4122: qgen: fix bitrotted cluster unit tests
There's a small set of pytest-style tests and associated conftest for
testing some of the cluster-related test infrastructure Python objects.
Going forward, I want unit tests for the query generator to be run as
part of patch acceptance (CI isn't necessary at this time).

This patch fixes a bitrotted test and moves the tests into the
tests-for-qgen directory. The moves were performed thusly:

 $ git mv conftest.py tests/
 $ git mv cluster_tests.py tests/test_cluster.py

Change-Id: I3e855e265ae245ebe3691d077284ac5761909e00
Reviewed-on: http://gerrit.cloudera.org:8080/4404
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins