IMPALA-4138: Fix AcquireState() for batches that change capacity
If MarkAtCapacity() is called on a row batch, it is difficult to call
AcquireState() on that batch because tuple_ptrs_size_ is not accessible
to initialise the destination batch - this is usually calculated from
capacity(), but that value is wrong for these purposes after
MarkAtCapacity().

Add RowBatch::InitialCapacity() to return the initial capacity value of
the batch.

Add row-batch-test to add initial coverage of AcquireState() API.

Change-Id: I6ceca53c406b05cd04b7d95a4f9f2eec7bc127f5
Reviewed-on: http://gerrit.cloudera.org:8080/4428
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins